+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### Changes
 
 - Added the minecraft version to the resulting mod jar
+- Added the custom query request handler